Return element and move util into separate module
